last update

+++ Project.Core/Features/Exams/Commands/Models/GradeExamCommand.cs
@@ -9,17 +9,17 @@ namespace Project.Core.Features.Exams.Commands.Models
     public class GradeAnswerDto
     {
         public int StudentAnswerId { get; set; }
-        public int PointsEarned { get; set; }
+        public decimal PointsEarned { get; set; } // ???? ??????? ???????
         public bool IsCorrect { get; set; }
-        public string? Feedback { get; set; }
+        public string? Feedback { get; set; } // ??????? ??????
     }
 
     public class GradeExamResponse
     {
         public int StudentExamResultId { get; set; }
-        public int PreviousTotalScore { get; set; }
-        public int NewTotalScore { get; set; }
-        public int PointsFromManualGrading { get; set; }
+        public decimal PreviousTotalScore { get; set; } // ???? ??????? ???????
+        public decimal NewTotalScore { get; set; } // ???? ??????? ???????
+        public decimal PointsFromManualGrading { get; set; } // ???? ??????? ???????
         public string Message { get; set; } = null!;
     }
 }
